+++ docs/changelog.rst
@@ -6,9 +6,12 @@ Unreleased
 
 * Add the option to only show progress on the dashboard. (`#107`_)
 * Import escape directly from markupsafe, instead of from flask. (`#106`_)
+* Insights now also work when using the ``forkserver`` and ``spawn`` start methods. (`#104`_)
+* When using insights on Windows the arguments of the top 5 longest tasks are now available as well.
 
 .. _#108: https://github.com/sybrenjansen/mpire/pull/107
 .. _#107: https://github.com/sybrenjansen/mpire/issues/106
+.. _#104: https://github.com/sybrenjansen/mpire/issues/104
 
 
 2.8.1